refactor: FSD Phase 6g — コマンドパレット・サイドバーのフックを widgets/app-shell へ移す (FRESTYLE-173)
widgets/app-shell の UI(CommandPalette / Header)だけが使うフックと定数を
widgets/app-shell の中へ移す。

- useCommandPalette → widgets/app-shell/model
- commandPaletteItems → widgets/app-shell/config
- useSidebar → widgets/app-shell/model(ログアウト操作を含む)

useSidebar は entities/user(clearAuth / AuthRepository)を使うが、
widgets → entities は許可された方向。widget 内部の参照は相対に統一。

これで constants は空になり削除。hooks は AI チャット系 + useAuth のみ、
components は MessageInput のみ(すべて Phase 6h で ask-ai + store とセット)。

既存テストの assert は変更していない。テスト件数 1184 で不変。
