+++ frontend/src/widgets/app-shell/config/commandPaletteItems.ts
