+++ frontend/src/widgets/app-shell/model/__tests__/useSidebar.test.ts
