+++ frontend/src/widgets/app-shell/model/useSidebar.ts
